+++ CHANGELOG.md
@@ -7,8 +7,8 @@ The version format is MCMAJOR.MCMINOR.MCPATCH-MAJOR.MINOR.PATCH, eg, 1.18.1-1.2.
 
 ## [UNRELEASED]
 ### Added
-- Added global loot modifier types `lapislib:replace_item` and `lapislib:remove_item`
-- Added configuration gui
+- Added global loot modifier types `lapislib:replace_item`, `lapislib:remove_item`, and `lapislib:add_entry`
+- Added configuration gui api
 
 ## [1.18.2-1.0.2a]
 ### Fixed